+++ packages/salesforcedx-vscode-core/src/commands/auth/forceAuthDevHub.ts
@@ -20,13 +20,16 @@ import {
 
 import { getRootWorkspacePath } from '../../util';
 
-import { ConfigFile, OrgConfigProperties } from '@salesforce/core';
+import { ConfigFile } from '@salesforce/core';
 import { isNullOrUndefined } from '@salesforce/salesforcedx-utils-vscode/out/src/helpers';
 import { ContinueResponse } from '@salesforce/salesforcedx-utils-vscode/out/src/types';
 import { homedir } from 'os';
 import * as vscode from 'vscode';
 import { CLI } from '../../constants';
-import { SFDX_CONFIG_FILE } from '../../constants';
+import {
+  DEFAULT_DEV_HUB_USERNAME_KEY,
+  SFDX_CONFIG_FILE
+} from '../../constants';
 import { nls } from '../../messages';
 import { isDemoMode } from '../../modes/demo-mode';
 import { isSFDXContainerMode } from '../../util';
@@ -107,7 +110,7 @@ export class ForceAuthDevHubExecutor extends SfdxCommandletExecutor<{}> {
       filename: SFDX_CONFIG_FILE
     });
 
-    globalConfig.set(OrgConfigProperties.TARGET_DEV_HUB, newUsername);
+    globalConfig.set(DEFAULT_DEV_HUB_USERNAME_KEY, newUsername);
     await globalConfig.write();
   }
 }