+++ packages/salesforcedx-vscode/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 55.7.0 - July 27, 2022
+
+## Fixed
+
+#### docs, #### salesforcedx-vscode-apex
+
+- We updated some Java installation links in our docs. ([PR #4305](https://github.com/forcedotcom/salesforcedx-vscode/pull/4305))
+
 # 55.6.0 - July 21, 2022
 
 ## Fixed